// src/types/models.ts
export type UserRole =
  | "super_admin"
  | "org_owner"
  | "director"
  | "medical_director"
  | "admin"
  | "dept_head"
  | "doctor"
  | "nurse"
  | "receptionist"
  | "pharmacist"
  | "cashier"
  | "cashier_agro"
  | "cashier_vet"
  | "accountant"
  | "hr_manager"
  | "storekeeper"
  | "lab_technician"
  | "radiologist"
  | "ambulance_staff"
  | "case_manager"
  | "social_worker"
  | "counselor"
  | "legal_officer"
  | "insurance_officer"
  | "it_manager"
  | "patient"
  | "auditor"
  | "cfo"
  | "finance_manager"
  | "procurement_officer"
  | "warehouse_manager"
  | "biomedical_engineer"
  | "compliance_officer"
  | "billing_officer"
  | "ot_coordinator"
  | "ward_manager"
  | "quality_officer";